Codes

+++ author-SANCHAY JOSHI(2).c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 
 int main()
-{
-	unsigned long var = 2147483648;
+{ 
+	long long int var = 2147483648;
 	printf("Sanchay\n\n");
-	printf("Value of var = %lu\n",var);
+	printf("Value of var = %lld\n",var);
 	return 0;
 }
